+++ swagger_parser/lib/src/generator/config/generator_config.dart
@@ -1,3 +1,4 @@
+import 'package:swagger_parser/src/generator/model/field_parser.dart';
 import 'package:swagger_parser/src/generator/model/json_serializer.dart';
 import 'package:swagger_parser/src/generator/model/programming_language.dart';
 import 'package:swagger_parser/src/parser/model/replacement_rule.dart';
@@ -32,6 +33,8 @@ class GeneratorConfig {
     this.mergeOutputs = false,
     this.includeIfNull = false,
     this.useFlutterCompute = false,
+    this.generateUrlsConstants = false,
+    this.fieldParsers = const [],
   });
 
   /// Optional. Set API name for folder and export file or merged output file
@@ -154,4 +157,11 @@ class GeneratorConfig {
   /// Optional. Set `true` to generate Retrofit clients with `Parser.FlutterCompute`
   /// and serialize/deserialize top-level functions for isolate-based multithreading.
   final bool useFlutterCompute;
+
+  /// DART/FLUTTER ONLY
+  /// Optional. Set `true` to generate URL constants for all endpoints.
+  final bool generateUrlsConstants;
+
+  /// {@macro field_parsers}
+  final List<FieldParser> fieldParsers;
 }