+++ swagger_parser/lib/src/parser/config/parser_config.dart
@@ -13,9 +13,11 @@ class ParserConfig {
     this.enumsParentPrefix = true,
     this.skippedParameters = const <String>[],
     this.replacementRules = const [],
+    this.replacementRulesForRawSchema = const [],
     this.useXNullable = false,
     this.excludeTags = const <String>[],
     this.includeTags = const <String>[],
+    this.includePaths,
     this.fallbackClient = 'fallback',
     this.inferRequiredFromNullable = false,
   });
@@ -55,6 +57,9 @@ class ParserConfig {
   /// All rules are applied in order.
   final List<ReplacementRule> replacementRules;
 
+  /// {@macro replacement_rules_for_raw_schema}
+  final List<ReplacementRule> replacementRulesForRawSchema;
+
   /// Does the Schema use x-nullable to indicate nullable fields
   /// Only used for OpenApi v2
   final bool useXNullable;
@@ -69,6 +74,9 @@ class ParserConfig {
   /// If set, only endpoints with these tags will be included in the generated clients.
   /// **NOTE: This will override the [excludeTags] if set.**
   final List<String> includeTags;
+
+  ///{@macro include_paths}
+  final List<String>? includePaths;
 
   /// DART ONLY
   /// Optional. Fallback client name for endpoints without tags.