+++ swagger_parser/test/e2e/tests/api_path_preservation/generate_test.dart
@@ -1,0 +1,14 @@
+import 'package:swagger_parser/swagger_parser.dart';
+
+Future<void> main() async {
+  final config = SWPConfig(
+    outputDirectory: 'test/e2e/tests/api_path_preservation/generated_files',
+    schemaPath: 'test/e2e/tests/api_path_preservation/openapi.yaml',
+    jsonSerializer: JsonSerializer.freezed,
+    putClientsInFolder: true,
+  );
+
+  final processor = GenProcessor(config);
+  await processor.generateFiles();
+  print('Files generated successfully!');
+}